% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.radiusDate}
\alias{as.radiusDate}
\title{Format Radius style date}
\usage{
as.radiusDate(date)
}
\arguments{
\item{date}{A date}
}
\value{
A character string
}
\description{
Format Radius style date
}
\examples{
as.radiusDate("2024-07-31")
}
