% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.rawFileName}
\alias{as.rawFileName}
\title{Format Radius style file name}
\usage{
as.rawFileName(root, date)
}
\arguments{
\item{root}{Raw data file root}

\item{date}{A date}
}
\value{
A character string
}
\description{
Format Radius style file name
}
\examples{
as.rawFileName("Students Export", "2024-07-31")
}
