% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.rawFilePath}
\alias{as.rawFilePath}
\title{Title}
\usage{
as.rawFilePath(dir, root, date)
}
\arguments{
\item{dir}{Raw data directory path}

\item{root}{Raw data file root}

\item{date}{A date}
}
\description{
Title
}
