% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{attendanceCheck}
\alias{attendanceCheck}
\title{Check which students have not attended}
\usage{
attendanceCheck(allowedBdays = 5)
}
\arguments{
\item{allowedBdays}{Minimum number of days not attended to flag}
}
\value{
A data frame
}
\description{
Check which students have not attended
}
\examples{
attendanceCheck()
attendanceCheck(allowedBdays = 10)
}
