% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnco-env-get.R
\name{cacheDir}
\alias{cacheDir}
\title{Get cache directory}
\usage{
cacheDir()
}
\value{
A character string
}
\description{
By default, this directory is set to \code{"mnco-cache"}. It can be changed
using \code{\link[=setCacheDir]{setCacheDir()}}.
}
\examples{
cacheDir()
}
