% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{createTextMessageFiles}
\alias{createTextMessageFiles}
\title{Format and store attendance text messages}
\usage{
createTextMessageFiles(flaggedStudents, date = Sys.Date())
}
\arguments{
\item{flaggedStudents}{Data frame of students}

\item{date}{Date for file name}
}
\value{
Character string containing the name of the file
}
\description{
The text messages is stored in .txt files and named using the date and
account name. All students under a single account are put into the
same text message.
}
