% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{edit_variable}
\alias{edit_variable}
\title{Edit a variable in the settings file}
\usage{
edit_variable(variable_name, value)
}
\arguments{
\item{variable_name}{Name of the variable}

\item{value}{Value being written to the variable}
}
\description{
Edit a variable in the settings file
}
