% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{expGradedifPest}
\alias{expGradedifPest}
\title{Plot grade difference vs. point estimate}
\usage{
expGradedifPest()
}
\value{
A ggplot object
}
\description{
Plot grade difference vs. point estimate
}
\examples{
# write later
}
