% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{generateMap}
\alias{generateMap}
\title{Create map of ???}
\usage{
generateMap(useCache = TRUE)
}
\arguments{
\item{useCache}{\code{logical} indicating to use the cache file}
}
\value{
A data frame
}
\description{
Create map of ???
}
\examples{
# write later
}
