% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data_3-tidy.R
\name{get_raw_na_cols}
\alias{get_raw_na_cols}
\title{Get NA column names}
\usage{
get_raw_na_cols(date = Sys.Date())
}
\arguments{
\item{date}{Date to read data for}
}
\value{
A vector of column names
}
\description{
Get NA column names
}
\examples{
get_raw_na_cols()
}
