% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo-inactive.R
\name{getActiveStudents}
\alias{getActiveStudents}
\title{Get enrolled students that are not pseudo-inactive}
\usage{
getActiveStudents()
}
\value{
A data frame
}
\description{
Get enrolled students that are not pseudo-inactive
}
