% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{getAssessments}
\alias{getAssessments}
\title{Read Radius assessment data}
\usage{
getAssessments(date = Sys.Date(), ignoreMissing = FALSE)
}
\arguments{
\item{date}{Date to read data for}

\item{ignoreMissing}{\code{logical} indicating to return empty data frame
(instead of signaling an error) if data file not found}
}
\value{
A data frame
}
\description{
Read Radius assessment data
}
\examples{
getAssessments()
getAssessments(date = "2024-07-31")
}
