% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data_4-get.R
\name{getAttendanceData}
\alias{getAttendanceData}
\title{Title}
\usage{
getAttendanceData(get = FALSE, date = Sys.Date())
}
\arguments{
\item{get}{\code{logical} indicating to return attendance log file}

\item{date}{}
}
\value{
A data frame if \code{get} is \code{TRUE}, otherwise none (invisible \code{NULL})
}
\description{
Title
}
\examples{
getAttendanceData()
}
