% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{getAttendanceTrainingSet}
\alias{getAttendanceTrainingSet}
\title{Read training data for attendance}
\usage{
getAttendanceTrainingSet()
}
\value{
A data frame
}
\description{
Read training data for attendance
}
\examples{
getAttendanceTrainingSet()
}
