% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{getHistoricAssessments}
\alias{getHistoricAssessments}
\title{Title}
\usage{
getHistoricAssessments(Assessments_Prior_to_ = "8_5_2024")
}
\arguments{
\item{Assessments_Prior_to_}{}
}
\description{
Title
}
