% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{getMostRecentAssessments}
\alias{getMostRecentAssessments}
\title{Get most recent assessment}
\usage{
getMostRecentAssessments(Assessments_Prior_to_ = "8_5_2024")
}
\arguments{
\item{Assessments_Prior_to_}{Radius style date}
}
\value{
A data frame
}
\description{
Get most recent assessment
}
\examples{
getMostRecentAssessments()
getMostRecentAssessments(Assessments_Prior_to_ = "8_5_2024")
}
