% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data-get.R
\name{getPaymentData}
\alias{getPaymentData}
\title{Read and tidy Radius payment data}
\usage{
getPaymentData(date = Sys.Date(), ignoreMissing = F)
}
\arguments{
\item{date}{Date to read data for}

\item{ignoreMissing}{\code{logical} indicating to return empty data frame
(instead of signaling an error) if data file not found}
}
\value{
A data frame
}
\description{
Read and tidy Radius payment data
}
\examples{
getPaymentData()
}
