% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{getPricing}
\alias{getPricing}
\title{Title}
\usage{
getPricing(
  gradeRange = "Kindergarten to Pre-Algebra",
  contractLength = 7,
  sessionCount = 10,
  nstu = dim(getCenterData("student")[which(getCenterData("student")$Enrollment_Status ==
    "Enrolled"), ])[1]
)
}
\arguments{
\item{nstu}{}
}
\description{
Title
}
