% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{getPricingGrid}
\alias{getPricingGrid}
\title{Title}
\usage{
getPricingGrid(
  verbose = FALSE,
  nstu = dim(getCenterData("student")[which(getCenterData("student")$Enrollment_Status ==
    "Enrolled"), ])[1]
)
}
\arguments{
\item{nstu}{}
}
\description{
Title
}
