% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo-inactive.R
\name{getPseudoInactiveStudents}
\alias{getPseudoInactiveStudents}
\title{Get pseudo-inactive students from cache}
\usage{
getPseudoInactiveStudents()
}
\value{
A data frame
}
\description{
Get pseudo-inactive students from cache
}
