% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{getStudentRanking}
\alias{getStudentRanking}
\title{Get student ranking data}
\usage{
getStudentRanking(date = Sys.Date())
}
\arguments{
\item{date}{Date to calculate rankings for}
}
\value{
A data frame
}
\description{
Get student ranking data
}
\examples{
getStudentRanking()
}
