% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck-flag.R
\name{getSuppressedStudents}
\alias{getSuppressedStudents}
\title{Retrieve list of suppressed students}
\usage{
getSuppressedStudents()
}
\value{
A data frame
}
\description{
Retrieve list of suppressed students
}
\examples{
# write later
}
