% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{getTemplate}
\alias{getTemplate}
\title{Retrieve template from template cache file}
\usage{
getTemplate(createdDate = "10/19/2021 6:55:40 PM", name = NULL)
}
\arguments{
\item{createdDate}{POSIXlt object used as key for template}

\item{name}{Optional argument to get template by name}
}
\value{
A character string
}
\description{
Retrieve template from template cache file
}
\examples{
# write later
}
