% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{kablize}
\alias{kablize}
\title{Format data frame as LaTeX table for .Rmd knit}
\usage{
kablize(todisplay, headerWidth = 12)
}
\arguments{
\item{todisplay}{A data frame}
}
\value{
LaTeX talbe
}
\description{
Format data frame as LaTeX table for .Rmd knit
}
\examples{
# write later
}
