% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck.R
\name{needsDeck.assessment}
\alias{needsDeck.assessment}
\title{Needs deck based on assessment}
\usage{
needsDeck.assessment()
}
\value{
A \code{\link{character}} vector of students
}
\description{
Needs deck based on assessment
}
\examples{
needsDeckBasedOnAssessment()
}
