% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{needsDeckBasedOnAssessment}
\alias{needsDeckBasedOnAssessment}
\title{Needs deck based on assessment}
\usage{
needsDeckBasedOnAssessment(date = Sys.time())
}
\arguments{
\item{date}{Date to look at data for}
}
\value{
A \code{\link{character}} vector of students
}
\description{
Needs deck based on assessment
}
\examples{
needsDeckBasedOnAssessment()
}
