% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data_4-get.R
\name{patchJoin}
\alias{patchJoin}
\title{Join data frames and patch NA values}
\usage{
patchJoin(..., .by, first = FALSE)
}
\arguments{
\item{...}{Data frames or list of data frames}

\item{.by}{Character string or vector of key columns used to join}

\item{first}{If \code{TRUE}, only the shared column that appears first in \code{.by} is
used to join? Matching occurs between each pair of consecutive data frames
in the order they are passed in. If \code{FALSE}, all key columns in \code{.by} must
must exist in all data frames.}
}
\value{
A data frame
}
\description{
Join data frames and patch NA values
}
\examples{
stu <- getCenterData("student")
acc <- getCenterData("account")
pro <- getCenterData("progress")
patchJoin(stu, acc, .by = "Account_Id")
patchJoin(list(stu, acc), .by = "Account_Id")
patchJoin(stu, acc, pro, .by = c("Student", "Account_Id"), first = T)

}
\keyword{internal}
