% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotHighestCor}
\alias{plotHighestCor}
\title{Create plot of most correlated variables}
\usage{
plotHighestCor(dataset, removeOutliers = FALSE, oCI = 99)
}
\arguments{
\item{dataset}{A data frame}

\item{removeOutliers}{\code{logical} indicating to remove outliers}

\item{oCI}{???}
}
\value{
Two ggplot objects
}
\description{
Create plot of most correlated variables
}
\examples{
# write later
}
