% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{pluralizeNames}
\alias{pluralizeNames}
\title{Format character strings in sentence list form}
\usage{
pluralizeNames(...)
}
\arguments{
\item{...}{Character strings or vectors}
}
\value{
A character string
}
\description{
Format character strings in sentence list form
}
\examples{
pluralizeNames("Justin")
pluralizeNames("Justin", "Luke")
pluralizeNames("Justin", "Luke", "Pat")
}
