% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data_3-tidy.R
\name{print_raw_na_cols}
\alias{print_raw_na_cols}
\title{Prints names of NA columns in raw data files}
\usage{
print_raw_na_cols(date = Sys.Date())
}
\arguments{
\item{date}{Date to read data for}
}
\value{
None (invisible \code{NULL})
}
\description{
The column names are formatted for copy/paste into a vector in code
}
\examples{
print_raw_na_cols
}
