% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data_2-read.R
\name{radiusDate}
\alias{radiusDate}
\title{Format Radius style date}
\usage{
radiusDate(date)
}
\arguments{
\item{date}{A date}
}
\value{
A character string
}
\description{
Format Radius style date
}
