% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnco-env-get.R
\name{rawDataDir}
\alias{rawDataDir}
\title{Get raw data directory}
\usage{
rawDataDir()
}
\value{
A character string
}
\description{
By default, this directory is set to \code{"mnco-raw-data"}. It can be changed
using \code{\link[=setRawDataDir]{setRawDataDir()}}.
}
\examples{
rawDataDir()
}
