% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data_2-read.R
\name{rawFileName}
\alias{rawFileName}
\title{Format Radius style file name}
\usage{
rawFileName(root, date)
}
\arguments{
\item{root}{Raw data file root}

\item{date}{A date}
}
\value{
A character string
}
\description{
Format Radius style file name
}
