% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data_2-read.R
\name{rawFilePath}
\alias{rawFilePath}
\title{Format Radius data file path}
\usage{
rawFilePath(dir, root, date)
}
\arguments{
\item{dir}{Raw data directory path}

\item{root}{Raw data file root}

\item{date}{A date}
}
\value{
A character string
}
\description{
Format Radius data file path
}
