% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{regularizeScore}
\alias{regularizeScore}
\title{Regularize student ranking score}
\usage{
regularizeScore(dat, variableName, centerVal)
}
\arguments{
\item{dat}{Data frame of students with scores}

\item{variableName}{Variable to regularize on}

\item{centerVal}{???}
}
\value{
A data frame
}
\description{
Regularize student ranking score
}
\examples{
# write later
}
