% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck-flag.R
\name{removeDeckSuppression}
\alias{removeDeckSuppression}
\title{Remove deck warning suppression from student}
\usage{
removeDeckSuppression(students = NULL)
}
\arguments{
\item{student}{Name of student to remove}
}
\value{
None (invisible(\code{NULL}))
}
\description{
Remove deck warning suppression from student
}
\examples{
# write later
}
