% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{removeDifferentDurationStudent}
\alias{removeDifferentDurationStudent}
\title{Unassign non-default session length from student}
\usage{
removeDifferentDurationStudent(student)
}
\arguments{
\item{student}{Name of student to remove}
}
\value{
None (invisible \code{NULL})
}
\description{
See \code{\link[=addDifferentDurationStudent]{addDifferentDurationStudent()}} for more details.
}
\examples{
removeDifferentDurationStudent("John Doe")
}
