% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo-inactive.R
\name{removePseudoInactiveStudent}
\alias{removePseudoInactiveStudent}
\title{Remove pseudo-inactive student}
\usage{
removePseudoInactiveStudent(students)
}
\arguments{
\item{who}{Student to remove}
}
\value{
Whether student was removed successfully
}
\description{
Remove pseudo-inactive student
}
