% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data-tidy.R
\name{removeRawCols}
\alias{removeRawCols}
\title{Delete columns from data frame}
\usage{
removeRawCols(df, ..., test_na = F)
}
\arguments{
\item{df}{A data frame}

\item{...}{Character strings or vectors of columns to remove}

\item{test_na}{If \code{TRUE}, signal error if column to remove in \code{df} does
not contain only \code{NA} values.}
}
\value{
A data frame
}
\description{
Delete columns from data frame
}
\examples{
stu <- readRawData("student")
stu <- stu \%>\% removeRawCols("Scholarship", "Center")
}
