% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{retrieve_variable}
\alias{retrieve_variable}
\title{Retrieve a variable from the settings file}
\usage{
retrieve_variable(variable_name, is_numeric = TRUE)
}
\arguments{
\item{variable_name}{Name of the variable}

\item{is_numeric}{Boolean to determine how the variable is interpreted}
}
\value{
Value from the file
}
\description{
Retrieve a variable from the settings file
}
