% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{returnFromVacation}
\alias{returnFromVacation}
\title{Remove student from vacation file}
\usage{
returnFromVacation(studentName = NULL)
}
\arguments{
\item{who}{Student to remove}
}
\value{
None (invisible \code{NULL})
}
\description{
Remove student from vacation file
}
\examples{
returnFromVacation("John Doe")
}
