% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{returnStudentFromVacation}
\alias{returnStudentFromVacation}
\title{Remove student from vacation file}
\usage{
returnStudentFromVacation(who)
}
\arguments{
\item{who}{Student to remove}
}
\value{
None (invisible \code{NULL})
}
\description{
Remove student from vacation file
}
\examples{
returnStudentFromVacation("John Doe")
}
