% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data-save.R
\name{saveAllCenterData}
\alias{saveAllCenterData}
\title{Title}
\usage{
saveAllCenterData(
  startDate = as.Date("2020-01-01"),
  endDate = Sys.Date(),
  promptDelete = T
)
}
\arguments{
\item{promptDelete}{}
}
\description{
Title
}
