% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data_5-save.R
\name{saveCenterData}
\alias{saveCenterData}
\title{Store center data in .csv file}
\usage{
saveCenterData(date = Sys.Date(), silent = F)
}
\arguments{
\item{date}{Date to save data for}

\item{silent}{\code{logical}. Determines if messages should be printed.}

\item{ignoreMissing}{\code{logical} indicating if data should be saved if there
is missing data}
}
\value{
None (invisible \code{NULL})
}
\description{
Store center data in .csv file
}
\examples{
saveCenterData()
}
