% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{saveTemplates}
\alias{saveTemplates}
\title{Save template to cache file}
\usage{
saveTemplates()
}
\value{
None (invisible \code{NULL})
}
\description{
Save template to cache file
}
\examples{
# write later
}
