% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{sendOnVacation}
\alias{sendOnVacation}
\title{Send student on vacation}
\usage{
sendOnVacation(..., returnDate = NULL)
}
\arguments{
\item{returnDate}{Date of return}

\item{who}{Student name}
}
\value{
None (invisible \code{NULL})
}
\description{
Send student on vacation
}
\examples{
sendOnVacation("John Doe", returnDate = "2024-07-31")
}
