% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnco-env-set.R
\name{setCacheDir}
\alias{setCacheDir}
\title{Set cache directory}
\usage{
setCacheDir(path)
}
\arguments{
\item{path}{A directory path}
}
\value{
The previous cache directory path, invisibly.
}
\description{
Set cache directory
}
\examples{
setCacheDir("./Cache")
}
