% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnco-env-set.R
\name{setDownloadsDir}
\alias{setDownloadsDir}
\title{Set downloads directory}
\usage{
setDownloadsDir(path)
}
\arguments{
\item{path}{A directory path}
}
\value{
The previous downloads directory path, invisibly.
}
\description{
Set downloads directory
}
\examples{
setDownloadsDir("./Downloads")
}
