% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo-inactive.R
\name{setPseudoInactiveStudent}
\alias{setPseudoInactiveStudent}
\title{Set student as pseudo-inactive}
\usage{
setPseudoInactiveStudent(students)
}
\arguments{
\item{students}{Character string or vector}
}
\value{
Whether student was successfully added
}
\description{
Set student as pseudo-inactive
}
