% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnco-env-set.R
\name{setRawDataDir}
\alias{setRawDataDir}
\title{Set Radius data directory}
\usage{
setRawDataDir(path)
}
\arguments{
\item{path}{A directory path}
}
\value{
The previous raw data directory path, invisibly.
}
\description{
Set Radius data directory
}
\examples{
setRawDataDir("./Raw_Data")
}
