% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{setStudentsOnVacation}
\alias{setStudentsOnVacation}
\title{Add student to vacation file}
\usage{
setStudentsOnVacation(
  dat = data.frame(matrix(ncol = 2, nrow = 0, dimnames = list(NULL, c("Student",
    "returnDate"))))
)
}
\arguments{
\item{dat}{Data frame of students to add}
}
\value{
None (invisible \code{NULL})
}
\description{
Add student to vacation file
}
\examples{
setStudentsOnVacation()
}
