% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck.R
\name{setSuppressedStudents}
\alias{setSuppressedStudents}
\title{(to be deprecated) Save students to suppression file}
\usage{
setSuppressedStudents(
  dat = data.frame(matrix(ncol = 7, nrow = 0, dimnames = list(NULL, c("Student",
    "Skills_Currently_Assigned", "Pest", "Skills_Mastered", "Attendances", "creation",
    "expDate"))))
)
}
\arguments{
\item{dat}{Data frame of students}
}
\value{
None (invisible \code{NULL})
}
\description{
(to be deprecated) Save students to suppression file
}
