% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{showcaseRegularizeScore}
\alias{showcaseRegularizeScore}
\title{Plot regularized scores}
\usage{
showcaseRegularizeScore()
}
\value{
A ggplot object
}
\description{
Plot regularized scores
}
\examples{
# write later
}
