% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffler.R
\name{shuffle_blue_cards}
\alias{shuffle_blue_cards}
\title{Takes in and shuffles demo pdfs}
\usage{
shuffle_blue_cards(visual_pdf, mental_pdf)
}
\arguments{
\item{visual_pdf}{Variable for Visual Demo Pdf}

\item{mental_pdf}{Variable for Mental Demo Pdf}
}
\value{
None (invisible \code{NULL})
}
\description{
Takes in and shuffles demo pdfs
}
\examples{
shuffle_blue_cards("1.pdf","2.pdf")
}
