% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck-flag.R
\name{suppressDeckWarning}
\alias{suppressDeckWarning}
\title{Exclude student from deck warnings}
\usage{
suppressDeckWarning(
  students,
  duration = retrieve_variable("Deck_Warning_Duration"),
  reason = NA_character_
)
}
\arguments{
\item{duration}{Number of days}

\item{...}{Character string or vectors of students}
}
\value{
None (invisible \code{NULL})
}
\description{
Exclude student from deck warnings
}
\examples{
# write later
}
