% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{templatesNeedUpdated}
\alias{templatesNeedUpdated}
\title{Get list of templates to be updated}
\usage{
templatesNeedUpdated()
}
\value{
A data frame
}
\description{
Get list of templates to be updated
}
\examples{
# write later
}
