% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data-tidy.R
\name{tidyAssessments}
\alias{tidyAssessments}
\title{Tidy Radius assessment data}
\usage{
tidyAssessments(dat)
}
\arguments{
\item{dat}{Data frame to tidy}
}
\value{
A data frame
}
\description{
Tidy Radius assessment data
}
