% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data-tidy.R
\name{tidyRawData.account}
\alias{tidyRawData.account}
\title{Tidy Radius account data}
\usage{
tidyRawData.account(data)
}
\arguments{
\item{data}{Data frame to tidy}
}
\value{
A data frame
}
\description{
Tidy Radius account data
}
