% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data_3-tidy.R
\name{tidyRawData.enrollment}
\alias{tidyRawData.enrollment}
\title{Tidy Radius enrollment data}
\usage{
tidyRawData.enrollment(data)
}
\arguments{
\item{data}{Data frame to tidy}
}
\value{
A data frame
}
\description{
Tidy Radius enrollment data
}
