% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data_3-tidy.R
\name{tidyRawData.progress}
\alias{tidyRawData.progress}
\title{Tidy Radius progress data}
\usage{
tidyRawData.progress(data)
}
\arguments{
\item{data}{Data frame to tidy}
}
\value{
A data frame
}
\description{
Tidy Radius progress data
}
