% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data-tidy.R
\name{tidyRawData.student}
\alias{tidyRawData.student}
\title{Tidy Radius student data}
\usage{
tidyRawData.student(data)
}
\arguments{
\item{data}{Data frame to tidy}
}
\value{
A data frame
}
\description{
Tidy Radius student data
}
