% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.rawFilePath}
\alias{as.rawFilePath}
\title{Format Radius data file path}
\usage{
as.rawFilePath(dir, root, date)
}
\arguments{
\item{dir}{Raw data directory path}

\item{root}{Raw data file root}

\item{date}{A date}
}
\value{
A character string
}
\description{
Format Radius data file path
}
\examples{
as.rawFilePath(rawDataDir(), "Student Export", "2024-07-31")
}
