% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data-get.R
\name{getCenterData}
\alias{getCenterData}
\title{Get tidied data from Radius files}
\usage{
getCenterData(
  type = c("all", radiusFileRoots("types")),
  date = Sys.Date(),
  ignoreMissing = F
)
}
\arguments{
\item{type}{Which kind of Radius data to get}

\item{date}{What date to get data for.}

\item{ignoreMissing}{Whether to throw an error if file is not found}
}
\value{
A data frame
}
\description{
Get tidied data from Radius files
}
