% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data-get.R
\name{getCenterData}
\alias{getCenterData}
\title{Read and tidy Radius data}
\usage{
getCenterData(
  type = c("all", radiusFileRoots("types")),
  date = Sys.Date(),
  ignoreMissing = F
)
}
\arguments{
\item{type}{String of which Radius data to read}

\item{date}{Date to read data for}

\item{ignoreMissing}{\code{logical} indicating to return empty data frame
(instead of signaling an error) if data file not found}
}
\value{
A data frame
}
\description{
Read and tidy Radius data
}
\examples{
getCenterData()
getCenterData("student")
}
