% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{getHistoricAssessments}
\alias{getHistoricAssessments}
\title{Read historic assessment data}
\usage{
getHistoricAssessments(Assessments_Prior_to_ = "8_5_2024")
}
\arguments{
\item{Assessments_Prior_to_}{Radius style date}
}
\value{
A data frame
}
\description{
Read historic assessment data
}
\examples{
getHistoricAssessments()
getHistoricAssessments(Assessments_Prior_to_ = "7_31_2024")
}
