% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{getPricing}
\alias{getPricing}
\title{Calculate monthly payment amount}
\usage{
getPricing(
  gradeRange = "Kindergarten to Pre-Algebra",
  contractLength = retrieve_variable("Standard_Contract_Length"),
  sessionCount = retrieve_variable("Standard_Number_Sessions"),
  nstu = dim(getCenterData("student")[which(getCenterData("student")$Enrollment_Status ==
    "Enrolled"), ])[1]
)
}
\arguments{
\item{gradeRange}{A character string}

\item{contractLength}{Number of months}

\item{sessionCount}{Number of sessions per month}

\item{nstu}{Number of enrolled students}
}
\value{
A numeric
}
\description{
Calculate monthly payment amount
}
\examples{
# write later
}
