% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{getPricingGrid}
\alias{getPricingGrid}
\title{Create table of monthly amounts}
\usage{
getPricingGrid(
  verbose = FALSE,
  nstu = dim(getCenterData("student")[which(getCenterData("student")$Enrollment_Status ==
    "Enrolled"), ])[1]
)
}
\arguments{
\item{verbose}{Whether to calculate price per session}

\item{nstu}{Number of enrolled students}
}
\value{
A data frame
}
\description{
Create table of monthly amounts
}
\examples{
# write later
}
