% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{getStudentsOnVacation}
\alias{getStudentsOnVacation}
\title{Get students on vacation}
\usage{
getStudentsOnVacation()
}
\value{
A data frame
}
\description{
Get students on vacation
}
\examples{
getStudentsOnVacation()
}
