% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnco-data.R
\docType{data}
\name{radiusFileRoots}
\alias{radiusFileRoots}
\title{Radius file roots}
\format{
A character vector of file roots:
\describe{
\item{Students Export}{Student information}
\item{Account Export}{Account information}
\item{Current Batch Detail Export}{Trailing 30 days data on attendance
and skills mastered}
\item{Enrollment Report}{Enrollment information}
}
}
\usage{
radiusFileRoots
}
\description{
All Radius files that can be processed by mnco.
}
\keyword{datasets}
