% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data-read.R, R/readRawData.R
\name{readRawData}
\alias{readRawData}
\title{Read Radius raw data excel file}
\usage{
readRawData(x, type = NULL, date = Sys.Date())

readRawData(x, type = NULL, date = Sys.Date())
}
\arguments{
\item{x}{File path or directory.}

\item{type}{Which radius data to use.}

\item{date}{What date to use.}
}
\value{
A data frame

A data frame
}
\description{
Read Radius raw data excel file

Read Radius raw data excel file
}
