% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data-read.R
\name{readRawData}
\alias{readRawData}
\title{Read Radius data excel file}
\usage{
readRawData(type, date = Sys.Date())
}
\arguments{
\item{type}{Radius data file type}

\item{date}{Date to read data for}
}
\value{
A data frame
}
\description{
Read Radius data excel file
}
\examples{
readRawData("student")
}
