% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck.R
\name{removeDeckSuppression}
\alias{removeDeckSuppression}
\title{Remove deck warning suppression from student}
\usage{
removeDeckSuppression(
  studentRows = data.frame(matrix(ncol = 7, nrow = 0, dimnames = list(NULL, c("Student",
    "Skills_Currently_Assigned", "Pest", "Skills_Mastered", "Attendances", "creation",
    "expDate"))))
)
}
\arguments{
\item{studentRows}{Data frame of students}
}
\value{
None (invisible(\code{NULL}))
}
\description{
Remove deck warning suppression from student
}
\examples{
# write later
}
