% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{sendOnVacation}
\alias{sendOnVacation}
\title{Send student on vacation}
\usage{
sendOnVacation(who, returnDate = lubridate::rollforward(Sys.Date()))
}
\arguments{
\item{who}{Student name}

\item{returnDate}{Date of return}
}
\value{
None (invisible \code{NULL})
}
\description{
Send student on vacation
}
