% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{setStudentsOnVacation}
\alias{setStudentsOnVacation}
\title{Title}
\usage{
setStudentsOnVacation(
  dat = data.frame(matrix(ncol = 2, nrow = 0, dimnames = list(NULL, c("Student",
    "returnDate"))))
)
}
\arguments{
\item{dat}{}
}
\description{
Title
}
