% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck.R
\name{setSuppressedStudents}
\alias{setSuppressedStudents}
\title{Title}
\usage{
setSuppressedStudents(
  dat = data.frame(matrix(ncol = 7, nrow = 0, dimnames = list(NULL, c("Student",
    "Skills_Currently_Assigned", "Pest", "Skills_Mastered", "Attendances", "creation",
    "expDate"))))
)
}
\arguments{
\item{dat}{}
}
\description{
Title
}
