% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck.R
\name{suppressDeckWarning}
\alias{suppressDeckWarning}
\title{Exclude student from deck warnings}
\usage{
suppressDeckWarning(
  studentRows = data.frame(matrix(ncol = 5, nrow = 0, dimnames = list(NULL, c("Student",
    "Skills_Currently_Assigned", "Pest", "Skills_Mastered", "Attendances")))),
  durationDays = retrieve_variable("Deck_Warning_Duration")
)
}
\arguments{
\item{studentRows}{Data frame of students}

\item{durationDays}{Number of days}
}
\value{
None (invisible \code{NULL})
}
\description{
Exclude student from deck warnings
}
\examples{
# write later
}
