% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck.R
\name{suppressDeckWarning}
\alias{suppressDeckWarning}
\title{Exclude student from deck warnings}
\usage{
suppressDeckWarning(..., duration = retrieve_variable("Deck_Warning_Duration"))
}
\arguments{
\item{...}{Character string or vectors of students}

\item{duration}{Number of days}
}
\value{
None (invisible \code{NULL})
}
\description{
Exclude student from deck warnings
}
\examples{
# write later
}
