% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck.R
\name{suppressDeckWarning}
\alias{suppressDeckWarning}
\title{Title}
\usage{
suppressDeckWarning(
  studentRows = data.frame(matrix(ncol = 5, nrow = 0, dimnames = list(NULL, c("Student",
    "Skills_Currently_Assigned", "Pest", "Skills_Mastered", "Attendances")))),
  durationDays = 2
)
}
\arguments{
\item{durationDays}{}
}
\description{
Title
}
