% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-data-tidy.R
\name{tidyRawData}
\alias{tidyRawData}
\title{Tidy Radius data}
\usage{
tidyRawData(data, type)
}
\arguments{
\item{data}{Data frame to tidy}

\item{type}{Radius data file type}
}
\value{
A data frame
}
\description{
Tidy Radius data
}
